Safari-609.1.11.1/LayoutTests/fast/css/counters/getCounterValue-expected.txt

Test the Counter object returned by the getCounterValue method of CSSPrimitiveValues

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS counters[0].identifier is "section"
PASS counters[0].listStyle is "decimal"
PASS counters[0].separator is ""
PASS counters[1].identifier is "section"
PASS counters[1].listStyle is "decimal"
PASS counters[1].separator is ":"
PASS counters[2].identifier is "section"
PASS counters[2].listStyle is "lower-roman"
PASS counters[2].separator is ""
PASS counters[3].identifier is "section"
PASS counters[3].listStyle is "upper-roman"
PASS counters[3].separator is ","
PASS counters[4].identifier is "section"
PASS counters[4].listStyle is "none"
PASS counters[4].separator is ""
PASS successfullyParsed is true

TEST COMPLETE
